Relax domain in VectorUtils.safeCastToLong ()
.. since we otherwise waste the long.

+++ community/cypher/runtime-util/src/main/java/org/neo4j/cypher/operations/VectorUtils.java
@@ -175,7 +175,7 @@ final class VectorUtils {
     }
 
     static long safeCastToLong(double value) {
-        if (value < Integer.MIN_VALUE || value > Integer.MAX_VALUE) {
+        if (value < Long.MIN_VALUE || value > Long.MAX_VALUE) {
             throw numericValueOutOfRange(String.valueOf(value), "vector()");
         }
         return (long) value;
